requirements.txt
praw
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
streamlit
streamlit-elements
streamlit-option-menu
plotly